.Net Core & EF Core with Knockout & Swagger
This web application uses modern UI and RESTful API, and as long as interfaces are the same, could be easily swapped out with a different application. RESTful API written in loosely coupled manner to keep the maintenance easier, and unit tests included to cover the logic implemented. The application requires minimal configuration to start.